Fix fatal error calling can() on non-existent keys

+++ lib/Hash/Objectify.pm
@@ -64,6 +64,7 @@ our $AUTOLOAD;
 
 sub can {
     my ( $self, $key ) = @_;
+    return unless ref $self && exists $self->{$key};
     $self->$key; # install accessor if not installed
     return $self->SUPER::can($key);
 }
